third_party/blink/web_tests/platform/generic/http/tests/workers/worker-invalid-context-expected.txt

CONSOLE ERROR: new Worker() has failed to fetch script.
This test checks that the invalid context doesn't cause crashes.

DONE
